restyle login button on landing page
