+++ services/frontend-v3/src/components/layouts/landingLayout/LandingLayoutView.jsx
@@ -1,11 +1,10 @@
-import React from "react";
+import React, { useState } from "react";
+import PropTypes from "prop-types";
 import { Row, Col, Typography, Button } from "antd";
+import { LockFilled, UnlockFilled } from "@ant-design/icons";
 import { FormattedMessage } from "react-intl";
 import { useKeycloak } from "@react-keycloak/web";
 import AppLayout from "../appLayout/AppLayout";
-import backgroundOptionOne from "../../../assets/landing-1.svg";
-import backgroundOptionTwo from "../../../assets/landing-2.svg";
-import backgroundOptionThree from "../../../assets/landing-3.svg";
 import logo from "../../../assets/I-talent-logo.png";
 import "./LandingLayoutView.scss";
 
@@ -16,31 +15,25 @@ const { Text, Title } = Typography;
  *
  *  this component renders the landing page.
  */
-const LandingLayoutView = () => {
-  /**
-   * Random Picture Select
-   *
-   * select a random picture for landing page
-   */
-  const randomPictureSelect = () => {
-    const imageOptions = [
-      backgroundOptionOne,
-      backgroundOptionTwo,
-      backgroundOptionThree,
-    ];
-    const randomIndex = Math.floor(Math.random() * imageOptions.length);
-    return imageOptions[randomIndex];
-  };
-
+const LandingLayoutView = ({ backgroundImage }) => {
   const { keycloak } = useKeycloak();
+  const [hover, setHover] = useState(false);
+
+  /**
+   * Toggle hover State
+   *
+   */
+  const toggleHover = () => {
+    setHover(!hover);
+  };
 
   return (
     <AppLayout displaySideBar={false} displaySearch={false} displayLogo={false}>
       <h1 className="hidden">
         <FormattedMessage id="landing.login.and.enter" />
       </h1>
-      <Row justify="center" style={{ marginTop: "120px" }}>
-        <Col xs={22} md={10} lg={6} style={{ paddingTop: "60px" }}>
+      <Row justify="center" className="page-content">
+        <Col xs={22} md={10} lg={6}>
           <img src={logo} alt="I-Talent Logo" className="logo" />
           <Title level={1} className="title">
             <FormattedMessage id="landing.welcome" />
@@ -51,20 +44,30 @@ const LandingLayoutView = () => {
           <Text className="text" strong>
             <FormattedMessage id="landing.call.to.action" />
           </Text>
-          <Button type="primary" onClick={() => keycloak.login()} size="large">
-            <FormattedMessage id="landing.login.button" />
+          <Button
+            type="primary"
+            onClick={() => keycloak.login()}
+            size="large"
+            className="sign-in-button"
+            onMouseEnter={toggleHover}
+            onMouseLeave={toggleHover}
+            icon={hover ? <UnlockFilled /> : <LockFilled />}
+          >
+            <strong>
+              <FormattedMessage id="landing.login.button" />
+            </strong>
           </Button>
         </Col>
         <Col sm={24} md={10} className="landingPicture imageContainer">
-          <img
-            src={randomPictureSelect()}
-            alt="I-Talent Logo"
-            className="image"
-          />
+          <img src={backgroundImage} alt="I-Talent Logo" className="image" />
         </Col>
       </Row>
     </AppLayout>
   );
+};
+
+LandingLayoutView.propTypes = {
+  backgroundImage: PropTypes.string.isRequired,
 };
 
 export default LandingLayoutView;
